requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
requests
beautifulsoup4
openai>=1.0.0
schedule
passlib
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
trafilatura
lxml_html_clean
pytz